refactor: remove jotai
