+++ libs/frontend/application/shared/store/src/ui/index.ts
@@ -1,3 +1,2 @@
 export * from './inline-form.service'
 export * from './modal.service'
-export * from './modal.state'
